backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.0
requests==2.31.0
ollama==0.3.0
llama-index==0.10.68
llama-index-embeddings-ollama==0.3.0
faiss-cpu==1.8.0
sentence-transformers==3.0.0
reportlab==4.2.0
langdetect==1.0.9
opencv-python==4.10.0.84
paddlepaddle==2.6.1
paddleocr==2.8.1